Add abstraction for wamp transports
This commit contains the initial work required to create an
abstraction for wamp transports. All of the message code has
been cleaned up so that it can be used properly through an
API. It can be further refined into a message hierarchy but
it is a good start for now. Furthermore, the rawsocket code
has been refactored out of the session class and is now the
basis for a wamp_rawsocket_transport.